Split lib/fragments into collection-specific fragment files
