+++ packages/lesswrong/lib/collections/messages/fragments.ts
@@ -1,0 +1,15 @@
+import { registerFragment } from '../../vulcan-lib/fragments';
+
+registerFragment(`
+  fragment messageListFragment on Message {
+    _id
+    user {
+      ...UsersMinimumInfo
+    }
+    contents {
+      html
+    }
+    createdAt
+    conversationId
+  }
+`);
