+++ packages/lesswrong/lib/collections/revisions/fragments.ts
@@ -1,6 +1,40 @@
 import { registerFragment } from '../../vulcan-lib';
 
 registerFragment(`
+  fragment RevisionDisplay on Revision {
+    _id
+    version
+    updateType
+    editedAt
+    userId
+    html
+    wordCount
+    htmlHighlight
+    plaintextDescription
+  }
+`)
+
+
+
+registerFragment(`
+  fragment RevisionEdit on Revision {
+    _id
+    version
+    updateType
+    editedAt
+    userId
+    originalContents
+    html
+    markdown
+    draftJS
+    ckEditorMarkup
+    wordCount
+    htmlHighlight
+    plaintextDescription
+  }
+`)
+
+registerFragment(`
   fragment RevisionMetadata on Revision {
     _id
     version
